menu fixs + clone weapon updated + skill damage  updated

+++ TRP/Skill.cs
@@ -7,9 +7,9 @@ namespace TRP
     class Skill
     {
         public string  Name { get; set; }
-        public int Damage { get; set; }
+        public double Damage { get; set; }
 
-        public Skill(string Name, int Damage)
+        public Skill(string Name, double Damage)
         {
             this.Name = Name;
             this.Damage = Damage;
